pyproject.toml
[project]
name = "fastfast"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "python-fasthtml>=0.5.1",
    "fastapi>=0.114.0",
    "uvicorn>=0.30.6",
    "loguru>=0.7.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
